ChannelMergerNode may need check for deferred updating of output channels https://bugs.webkit.org/show_bug.cgi?id=108863
There can in rare cases be a slight delay before the output bus is updated
to the new number of channels because of tryLocks() in the context's
updating system. So need to check the channel number before processing.

Reviewed by Chris Rogers.

* Modules/webaudio/ChannelMergerNode.cpp:
(WebCore::ChannelMergerNode::ChannelMergerNode):
(WebCore::ChannelMergerNode::process):
(WebCore::ChannelMergerNode::checkNumberOfChannelsForInput):
* Modules/webaudio/ChannelMergerNode.h:
(ChannelMergerNode):


Identifier: 128717@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@143565 268f45cc-cd09-0410-ab3c-d52691b4dbfc