Expose a list of all reasons that qualify a RenderLayer to be composited https://bugs.webkit.org/show_bug.cgi?id=110505
Expose the various reasons that can qualify a RenderLayer to be composited
via a new bitmask returned by the reasonForCompositing() method on
RenderLayerCompositor. This method already existed and was used for logging
purposes, but the previous functionality is now accessed via the
logReasonForCompositing() method.

This will allow the the InspectorLayerTreeAgent to provide this information
to the front-end on a per-layer basis as requested by the front-end.

Reviewed by Simon Fraser.

No new tests.

* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::logLayerInfo):
(WebCore::RenderLayerCompositor::reasonsForCompositing):
(WebCore):
(WebCore::RenderLayerCompositor::logReasonsForCompositing):
* rendering/RenderLayerCompositor.h:
(RenderLayerCompositor):


Identifier: 128912@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@143787 268f45cc-cd09-0410-ab3c-d52691b4dbfc