SQLTransactionBackend::doCleanup() should not nullify its current statement and transaction error fields. https://bugs.webkit.org/show_bug.cgi?id=110826.
Reviewed by Geoffrey Garen.

The reason for this is because the frontend SQLTransaction may have
tasks that are already registered via callOnMainThread(), and may
invoke transaction state functions that access these objects in the
SQLTransactionBackend. This proves to be problematic if the database
was interrupted resulting in doCleanup() being called concurrently
while those states are executing.

Instead, we leave their clean up to the SQLTransactionBackend destructor
which is guaranteed to only be called when the SQLTransaction is also
destructing i.e. these objects are no longer needed.

No new tests.

* Modules/webdatabase/SQLTransactionBackend.cpp:
(WebCore::SQLTransactionBackend::doCleanup):



Identifier: 129223@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@144135 268f45cc-cd09-0410-ab3c-d52691b4dbfc