Remove unused member variable m_useLatchedEventNode from PlatformWheelEvent.h https://bugs.webkit.org/show_bug.cgi?id=107314
Reviewed by Darin Adler.

EventHandler::handleGestureScrollCore() was removed in http://trac.webkit.org/changeset/140177,
and this was the only place where the member variable |m_useLatchedEventNode| in
PlatformWheelEvent was mutated. This variable is no longer needed and so it should be removed.

No change in behavior, so no new tests needed.

* platform/PlatformWheelEvent.h:
(WebCore::PlatformWheelEvent::PlatformWheelEvent):
(WebCore::PlatformWheelEvent::useLatchedEventNode):
(PlatformWheelEvent):


Identifier: 129587@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@144531 268f45cc-cd09-0410-ab3c-d52691b4dbfc