[Mac] Get rid of a useless method from DOMHTMLInputElement https://bugs.webkit.org/show_bug.cgi?id=111279
Reviewed by Sam Weinig.

Source/WebKit/mac:

* DOM/WebDOMOperations.mm:
* DOM/WebDOMOperationsPrivate.h:
setValueForUser: is already privately exposed by DOMHTMLInputElement,
no need to have it twice.

Tools:

* DumpRenderTree/mac/TestRunnerMac.mm:
(TestRunner::setValueForUser):
The method _setValueForUser: was added for testing. This feature is already
exposed by DOMHTMLInputElement as a private API and it is how this code is exercised
in real use cases. Get rid of the method created for testing, use the real thing
instead.


Identifier: 129643@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@144596 268f45cc-cd09-0410-ab3c-d52691b4dbfc