[New Multicolumn] Make sure region styling works for columns inside regions. https://bugs.webkit.org/show_bug.cgi?id=111276
Reviewed by Sam Weinig.

Source/WebCore:

Test: fast/regions/region-styling-in-columns.html

* rendering/RenderFlowThread.cpp:
(WebCore::CurrentRenderFlowThreadMaintainer::CurrentRenderFlowThreadMaintainer):
(WebCore::CurrentRenderFlowThreadMaintainer::~CurrentRenderFlowThreadMaintainer):
* rendering/RenderFlowThread.h:
(CurrentRenderFlowThreadMaintainer):
Fix the CurrentRenderFlowThreadMaintainer so that it saves off the old render flow thread
and restores it when done. This effectively makes flow thread layout a push/pop stack,
since with in-flow threads, this is allowed. We assert that a nested flow thread is
in-flow to preserve the invariant that you can't nest CSS Regions flow thread.

* rendering/RenderRegion.cpp:
(WebCore::RenderRegion::computeChildrenStyleInRegion):
Add isInFlowRenderFlowThread() to the anonymous check to make sure a style gets made
for the nested flow thread.

LayoutTests:

* fast/regions/region-style-in-columns-expected.html: Added.
* fast/regions/region-style-in-columns.html: Added.



Identifier: 129678@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@144633 268f45cc-cd09-0410-ab3c-d52691b4dbfc