[chromium] DRT crashes in an Aura/Win build https://bugs.webkit.org/show_bug.cgi?id=111394
Reviewed by Tony Chang.

When building (and running) with use_aura=1, we assume that
we're using the default theme engine, but DRT creates the
DRTWin theme engine and overwrites the default, leading to crashes
in basically every test from the mismatch.

This fix gets us down to at least "lots of crashes" :).

* DumpRenderTree/chromium/TestShellWin.cpp:
(platformInit):

Identifier: 129828@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@144802 268f45cc-cd09-0410-ab3c-d52691b4dbfc