Inline min/maxInstrinsicLogicalWidth functions https://bugs.webkit.org/show_bug.cgi?id=111399
Reviewed by Tony Chang.

Turns out these functions don't need to be public and only
have one caller. This patch is in preparation for supporting
intrinsic widths on tables, which will need to pass in the
border and padding width as an argument.

* rendering/RenderBox.cpp:
(WebCore::RenderBox::fillAvailableMeasure):
(WebCore::RenderBox::computeIntrinsicLogicalWidthUsing):
(WebCore::RenderBox::computeLogicalWidthInRegionUsing):
* rendering/RenderBox.h:

Identifier: 129840@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@144816 268f45cc-cd09-0410-ab3c-d52691b4dbfc