Fix default background of a dragged image. https://bugs.webkit.org/show_bug.cgi?id=110512
Reviewed by Tony Chang.

.:

* ManualTests/drag-background-with-padding.html: Added.

Source/WebCore:

Transparent background color (particularly visible in the padding
of an element) was not being respected. The solution is to
initially fill the drag image with transparent instead of white,
so that any color background works correctly. This patch does not
affect Apple Mac or Win platforms, given that this part of the
code is #if guarded from those platforms.

Manual test added:
  ManualTests/drag-background-with-padding.html

Infrastructure for retrieving the drag image for pixel testing
does not yet exist.

* page/Frame.cpp:
(WebCore::Frame::nodeImage):


Identifier: 129863@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@144841 268f45cc-cd09-0410-ab3c-d52691b4dbfc