fast/css/hover-update.html is flaky https://bugs.webkit.org/show_bug.cgi?id=111592
Reviewed by Ryosuke Niwa.

Speculative fix for a flaky test. The test sometimes fails on slower machines (mostly debug).
Re-check the results in a loop for up to 200ms to see if the :hover pseudo style applies.

Also force a layout after changing the style, which should also trigger the :hover style.

* fast/css/hover-update.html:
* platform/chromium/TestExpectations:
* platform/efl/TestExpectations:
* platform/gtk/TestExpectations:
* platform/mac/TestExpectations:


Identifier: 129977@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@144963 268f45cc-cd09-0410-ab3c-d52691b4dbfc